API Changes, Sinusoidal Easing.
The API can actually be further simplified by deriving a generalised
function for calculating the symmetrical and inverse functions of
given easing transformers. This means that the only concrete code is
in the "in" declarations; everything else is achieved through `out`
or `inAndOut`.

Oh, also, sinusoidal easing. Would've been its own commit, but Git
couldn't separate that hunk. Boo.